requirements.txt
python-frontmatter==0.5.0
-e git://github.com/chdemko/pandoc-latex-environment.git#egg=master
click==7.1.2
